✏️ #154 fix: 타입 위치 정리
