+++ lib/hooks/useInfiniteGoalsQuery.ts
@@ -1,11 +1,8 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { Goals } from '../types';
+import { Goals } from '../types/todo';
 import baseFetch from '../api/baseFetch';
 
-const useInfiniteGoalsQuery = (
-  size: number = 3,
-  sortOrder?: 'oldest' | 'newest',
-) => {
+const useInfiniteGoalsQuery = (size: number = 3, sortOrder?: 'oldest' | 'newest') => {
   return useInfiniteQuery<Goals>({
     queryKey: ['goals', { size, sortOrder }],
     queryFn: ({ pageParam = null }) => {
@@ -16,7 +13,7 @@ const useInfiniteGoalsQuery = (
 
       return baseFetch(`/4-4-dev/goals?${params.toString()}`);
     },
-    initialPageParam: null,  // 첫 요청에는 cursor가 null
+    initialPageParam: null, // 첫 요청에는 cursor가 null
     getNextPageParam: (lastPage) => lastPage.nextCursor ?? null, // 다음 cursor 반환
     refetchOnMount: false,
     refetchOnWindowFocus: false,
