--- lib/types/todos.ts
@@ -1,28 +1,0 @@
-export interface GoalInTodo {
-  id: number;
-  title: string;
-}
-
-export interface Todo {
-  noteId: number | null;
-  done: boolean;
-  linkUrl: string | null;
-  fileUrl: string | null;
-  title: string;
-  id: number;
-  goal: GoalInTodo | null;
-  userId: number;
-  teamId: string;
-  updatedAt: string;
-  createdAt: string;
-}
-
-export interface GetTodosResponse {
-  todos: Todo[];
-  totalCount: number;
-  nextCursor?: number;
-}
-
-export interface GetTodosProgressResponse {
-  progress: number;
-}
